Refactor message signatures
