+++ crates/valence_protocol/src/packets/s2c.rs
@@ -22,6 +22,7 @@ use crate::LengthPrefixedArray;
 pub mod commands;
 pub mod declare_recipes;
 pub mod map_data;
+pub mod message_signature;
 pub mod particle;
 pub mod player_chat_message;
 pub mod player_info_update;
@@ -107,6 +108,7 @@ pub mod login {
 pub mod play {
     use commands::Node;
     pub use map_data::MapData;
+    pub use message_signature::MessageSignature;
     pub use particle::ParticleS2c;
     pub use player_chat_message::PlayerChatMessage;
     pub use player_info_update::PlayerInfoUpdate;
